Components toegevoegd

+++ ConfigMan/ConfigMan/Controllers/ComponentsController.cs
@@ -7,43 +7,93 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
-using ConfigMan;
+using ConfigMan.ViewModels;
+using System.Diagnostics.Contracts;
 using ConfigMan.ActionFilters;
+using System.Diagnostics;
 
 namespace ConfigMan.Controllers
 {
+
     [LogActionFilter]
     [HandleError]
 
     public class ComponentsController : Controller
     {
-        private DbEntities db = new DbEntities();
+        private readonly DbEntities db = new DbEntities();
+        private static bool ContractErrorOccurred = false;
+        private static string ErrorMessage = " ";
 
+        public ActionResult Menu()
+        {
+            ViewBag.SympaMsg = TempData["SympaMsg"];
+            return View();
+        }
+
+        //
         // GET: Components
+        //
         public ActionResult Index()
         {
-           
-            ViewBag.SympaMsg = TempData["SympaMsg"];
-            return View(db.Components.OrderBy(x => x.ComponentName).ToList());
+            if (ContractErrorOccurred)
+            {
+                ContractErrorOccurred = false;
+                ViewBag.SympaMsg = ErrorMessage;
+                ErrorMessage = " ";
+            }
+            else
+            {
+                ViewBag.SympaMsg = TempData["SympaMsg"];
+            }
+            List<Component> complist = db.Components.OrderBy(x => x.ComponentName).ToList();
+            List<ComponentVM> vmlist = new List<ComponentVM>();
+            foreach (Component c in complist)
+            {
+                ComponentVM VM = new ComponentVM();
+                VM.Fill(c);
+                vmlist.Add(VM);
+            }
+            return View(vmlist);
 
         }
 
+        //
         // GET: Components/Details/5
+        //
         public ActionResult Details(int? id)
         {
-            if (id == null)
+
+            Contract.ContractFailed += (Contract_ContractFailed);
+            Contract.Requires(id > 0, "Contract Failed!");
+
+            if (!ContractErrorOccurred)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                Component component = db.Components.Find(id);
+                if (component == null)
+                {
+                    TempData["SympaMsg"] = "*** ERROR *** ID " + id.ToString() + " not found in database.";
+                }
+                else
+                {
+                    ComponentVM componentVM = new ComponentVM();
+                    componentVM.Fill(component);
+                    return View(componentVM);
+                }
             }
-            Component component = db.Components.Find(id);
-            if (component == null)
-            {
-                return HttpNotFound();
-            }
-            return View(component);
+            return RedirectToAction("Index");
         }
 
+        private void Contract_ContractFailed(object sender, ContractFailedEventArgs e)
+        {
+            ErrorMessage = "*** ERROR *** Selected component id is invalid.";
+            ContractErrorOccurred = true;
+            e.SetHandled();
+            return;
+        }
+
+        //
         // GET: Components/Create
+        //
         public ActionResult Create()
         {
             return View();
@@ -54,40 +104,55 @@ namespace ConfigMan.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ComponentID,ComponentName")] Component component)
+        public ActionResult Create([Bind(Include = "ComponentID,VendorID,ComponentName")] ComponentVM componentVM)
         {
             if (ModelState.IsValid)
             {
+                Component component = new Component();
+                component.Fill(componentVM);
                 bool addfailed = false;
                 db.Components.Add(component);
-                try {
+                try
+                {
                     db.SaveChanges();
-                    TempData["SympaMsg"] = component.ComponentName + " succesfully added";
+                    TempData["SympaMsg"] = "Component " + component.ComponentName + " succesfully added.";
                 }
-                catch (DbUpdateException) {
+                catch (DbUpdateException)
+                {
                     addfailed = true;
-                    ViewBag.SympaMsg = component.ComponentName + " already exists, use update function";
+                    ViewBag.SympaMsg = "Component " + component.ComponentName + " already exists, use update function.";
                 }
                 if (addfailed) { return View(); }
                 else { return RedirectToAction("Index"); }
             }
-            return View(component);
+
+            return View(componentVM);
         }
 
+        //
         // GET: Components/Edit/5
+        //
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            Contract.Requires((id != null) && (id > 0));
+            Contract.ContractFailed += (Contract_ContractFailed);
+
+            if (!ContractErrorOccurred)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                Component component = db.Components.Find(id);
+                if (component == null)
+                {
+                    TempData["SympaMsg"] = "*** ERROR *** ID " + id.ToString() + " not found in database.";
+                }
+                else
+                {
+                    ComponentVM componentVM = new ComponentVM();
+                    componentVM.Fill(component);
+                    return View(componentVM);
+                }
             }
-            Component component = db.Components.Find(id);
-            if (component == null)
-            {
-                return HttpNotFound();
-            }
-            component.ComponentName = component.ComponentName.Trim();
-            return View(component);
+            return RedirectToAction("Index");
+
         }
 
         // POST: Components/Edit/5
@@ -95,43 +160,98 @@ namespace ConfigMan.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ComponentID,ComponentName")] Component component)
+        public ActionResult Edit([Bind(Include = "ComponentID,VendorID,ComponentName")] ComponentVM componentVM)
         {
             if (ModelState.IsValid)
             {
+                Component component = new Component();
+                component.Fill(componentVM);
                 db.Entry(component).State = EntityState.Modified;
                 db.SaveChanges();
-                TempData["SympaMsg"] = component.ComponentName + " succesfully updated";
+                TempData["SympaMsg"] = "Component " + component.ComponentName + " succesfully updated.";
                 return RedirectToAction("Index");
             }
-            return View(component);
+            else
+            {
+                TempData["SympaMsg"] = "Update failed for component " + componentVM.ComponentName;
+                return View(componentVM);
+            }
+
         }
 
+        //
         // GET: Components/Delete/5
+        //
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            Contract.Requires((id != null) && (id > 0));
+            Contract.ContractFailed += (Contract_ContractFailed);
+
+            if (!ContractErrorOccurred)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                Component component = db.Components.Find(id);
+                if (component == null)
+                {
+                    TempData["SympaMsg"] = "*** ERROR *** ID " + id.ToString() + " not found in database.";
+                }
+                else
+                {
+                    ComponentVM componentVM = new ComponentVM();
+                    componentVM.Fill(component);
+                    return View(componentVM);
+                }
             }
-            Component component = db.Components.Find(id);
-            if (component == null)
-            {
-                return HttpNotFound();
-            }
-            return View(component);
+            return RedirectToAction("Index");
+
         }
+
 
         // POST: Components/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Component component = db.Components.Find(id);
-            db.Components.Remove(component);
-            db.SaveChanges();
-            TempData["SympaMsg"] = component.ComponentName + " succesfully deleted";
+            Contract.Requires(id > 0);
+            Contract.ContractFailed += (Contract_ContractFailed);
+            if (!ContractErrorOccurred)
+            {
+                Component component = db.Components.Find(id);
+                db.Components.Remove(component);
+                db.SaveChanges();
+                TempData["SympaMsg"] = "Component " + component.ComponentName + " succesfully deleted.";
+            }
             return RedirectToAction("Index");
+        }
+
+        public ActionResult Report01()
+        {
+            if (ContractErrorOccurred)
+            {
+                ContractErrorOccurred = false;
+                ViewBag.SympaMsg = ErrorMessage;
+                ErrorMessage = " ";
+            }
+            else
+            {
+                ViewBag.SympaMsg = TempData["SympaMsg"];
+            }
+            var query = from c in db.Components
+                        where !(from i in db.Installations
+                                select i.ComponentID)
+                               .Contains(c.ComponentID)
+                        orderby c.ComponentName
+                        select new ComponentVM
+                        {
+                            ComponentID = c.ComponentID,
+                            ComponentName = c.ComponentName,
+                            VendorID = c.VendorID
+                            
+                        };
+
+            List<ComponentVM> complist = query.ToList();
+
+            return View(complist);
+
         }
 
         protected override void Dispose(bool disposing)
@@ -142,5 +262,8 @@ namespace ConfigMan.Controllers
             }
             base.Dispose(disposing);
         }
+
     }
+
+
 }
